% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual.R
\name{render_package_manual}
\alias{render_package_manual}
\title{Generate HTML reference manual}
\usage{
render_package_manual(package, outdir = ".")
}
\arguments{
\item{package}{name of the package}

\item{outdir}{where to put the html file}
}
\description{
Renders complete package reference manual in HTML format.
}
\examples{
htmlfile <- render_package_manual('nnet', tempdir())
utils::browseURL(htmlfile)
}
